Some Refactor
