+++ clocking-server/src/shutdown.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ShutdownReason {
-    SIGINT,
-    SIGTERM,
+    Sigint,
+    Sigterm,
     SignalChannelClosed,
 }
